+++ tools/tpm2_verifysignature.c
@@ -201,7 +201,7 @@ static bool on_option(char key, char *value) {
 	case 'c':
 	    ctx.context_arg = value;
 	    break;
-	case 'g': {
+	case 'G': {
 		ctx.halg = tpm2_alg_util_from_optarg(value);
 		if (ctx.halg == TPM2_ALG_ERROR) {
 			LOG_ERR("Unable to convert algorithm, got: \"%s\"", value);
@@ -260,7 +260,7 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
 
     const struct option topts[] = {
             { "digest",       required_argument, NULL, 'D' },
-            { "halg",         required_argument, NULL, 'g' },
+            { "halg",         required_argument, NULL, 'G' },
             { "message",      required_argument, NULL, 'm' },
             { "input-format", required_argument, NULL, 'i' },
             { "sig",          required_argument, NULL, 's' },
@@ -269,7 +269,7 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
     };
 
 
-    *opts = tpm2_options_new("g:m:D:i:s:t:c:", ARRAY_LEN(topts), topts,
+    *opts = tpm2_options_new("G:m:D:i:s:t:c:", ARRAY_LEN(topts), topts,
                              on_option, NULL, TPM2_OPTIONS_SHOW_USAGE);
 
     return *opts != NULL;